third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/intrinsic-auto-repeat/row-auto-repeat-auto-009-expected.txt

This is a testharness.js-based test.
[FAIL] .masonry 1
  assert_equals: \n<div class="masonry" data-expected-width="300" data-expected-height="200">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
[FAIL] .masonry 2
  assert_equals: \n<div class="masonry" style="width: 200px; height: 100px;" data-expected-width="300" data-expected-height="200">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
[FAIL] .masonry 3
  assert_equals: \n<div class="masonry" style="width: min-content; height: min-content;" data-expected-width="300" data-expected-height="200">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
[FAIL] .masonry 4
  assert_equals: \n<div class="masonry" style="width: max-content; height: max-content;" data-expected-width="300" data-expected-height="200">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
[FAIL] .masonry 5
  assert_equals: \n<div class="masonry border" data-expected-width="320" data-expected-height="220">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
[FAIL] .masonry 6
  assert_equals: \n<div class="masonry border" style="width: 200px; height: 100px;" data-expected-width="320" data-expected-height="220">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
[FAIL] .masonry 7
  assert_equals: \n<div class="masonry border" style="width: min-content; height: min-content;" data-expected-width="320" data-expected-height="220">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
[FAIL] .masonry 8
  assert_equals: \n<div class="masonry border" style="width: max-content; height: max-content;" data-expected-width="320" data-expected-height="220">\n  <div class="item" data-offset-x="0" data-offset-y="112.5" data-expected-width="300" data-expected-height="50"></div>\n</div>\noffsetTop expected 112.5 but got 0
Harness: the test ran to completion.
